Metadata element writer raises exception when value is not supported.
Unsupported metadata values include Hash, Array, and Ezid::Metadata instances.
This is not a backward-incompatible change insofar as using them would trigger
unexpected errors (see #69) or not work as expected.

Although the API of Ezid::Metadata is marked private, it's worth noting
here that the initializer has been updated to be compatible with
superclass methods such as `#dup` that depend on the signature of
of the superclass initializer.